Initial commit
Add files related to only client side that are also present in repository https://github.com/OleksiyBechke/Melovi. In that repo you need to deploy the project with .NET while the it  works without the framework. There's no need in such complication, so I decided to create a new repo in order to simplify the installation proces and add here changes I did.